feat: merged claim form & existing claim form
